fix(identity)_: no need to double-check ENS and should return an error (#6474)
Potentially fixes https://github.com/status-im/status-desktop/issues/16224

This condition is already being checked by `ValidateDisplayName` just below.
The fact that we check it early and return without an error when we try to set an ENS can lead to an false state where the settings get modified from the client by accident.