fix(swap)!: remove RouterBuildTransactionsParams and update related functions
The main reason, based on the response we got from the paraswap team, for build tx error that was happening
not always but more or less often, was missed slippage param. In the past slippage parameter was provided at
the later stage, when the tx was about to be built for real, but now, since we fetch the tx data from paraswap api
at the moment we do the gas estimation, the slippage param is necessary earlier. That's why we have to provide
it at the moment of fetching the route and also clients need to re-fetch the route on every slippage change.

What's changed:
- Removed the `RouterBuildTransactionsParams` type and its usage across the codebase.
- Updated `BuildTransactionsFromRoute` function signatures to use `uuid` instead of the removed type.
- Adjusted related database operations and tests to reflect the removal of build input parameters.
- `SwapParaswapProcessor` updated to use slippage param appropriately, reduced calls to paraswap api, by storing tx data