fix(cryptocompare)_: suppress a crypto compare down state when fetching prices for non existing tokens
When fetching prices (either pricemulti or pricemultifull) from cryptocompare an error is returned if
cryptocompare doesn't know for the symbols the app asks for. Because of that error the app shows
cryptocompare as a market provider unavailable and because of that some features are not available
on the app start.

Checking the cryptocompare api an error is returned only if none of symbols from the request is supported
by cryptocompare, but if at least a single symbol is supported by cryptocompare its details will be returned
in the response. Based on that the issue is solved by adding ETH symbol to the request since we know for
sure that the info for it exists and that's how we overcame the issue.