fix(wallet)_: fee calculation improvements for status sepolia chain
Since the Status chain is gasless the old implementation  of `isEIP1559Compatible` function
(that was checking the base fee of `eth_feeHistory` in order to dynamically determine if the chain
is eip1559 compliant or not) was identifying it as a none eip1559 compliant chain, while it is for real.
These changes fix that by hardcodding the the flag for Status Sepolia chain.