chore_: cleaning the accounts manager
The following functions have been removed from the account package:
- `Recover`
- `CanRecover`
- `Sign`

and the rest of the code update to use the same functions from the personal api package,
where we already have those functions.

`MockedPersonalAPI` within the personal package was added as the easiest way of fixing
tests (basically it does what was done in the `AccountManagerMock`, cause any other approach
requires a lot of changes in communities/messenger tests.