perf(envelopes)_: only cache the hash and Processed status instead of whole envelope
Fixes #6523

Changes the envelopeCache to store a bool instead of a full envelope. That bool represents whether the envelope has been Processed.
This is sufficient info for the code to process envelopes the same way as before.
Only one test was affected by this. Other tests were just cleaned up and made simpler by that change, but still test as much.